+++ client/src/game/layers/layer.ts
@@ -133,21 +133,21 @@ export class Layer {
                     continue;
                 shape.drawAuras(ctx);
             }
-            this.shapes.forEach(shape => {
-                if (shape.options.has("skipDraw") && shape.options.get("skipDraw")) return;
-                if (shape.labels.length === 0 && gameStore.filterNoLabel) return;
+            for (const shape of this.shapes) {
+                if (shape.options.has("skipDraw") && shape.options.get("skipDraw")) continue;
+                if (shape.labels.length === 0 && gameStore.filterNoLabel) continue;
                 if (
                     shape.labels.length &&
                     gameStore.labelFilters.length &&
                     !shape.labels.some(l => gameStore.labelFilters.includes(l.uuid))
                 )
-                    return;
-                if (layerManager.getLayer() === undefined) return;
-                if (!shape.visibleInCanvas(state.canvas)) return;
+                    continue;
+                if (layerManager.getLayer() === undefined) continue;
+                if (!shape.visibleInCanvas(state.canvas)) continue;
                 if (state.name === "fow" && shape.visionObstruction && layerManager.getLayer()!.name !== state.name)
-                    return;
+                    continue;
                 shape.draw(ctx);
-            });
+            }
 
             if (this.selection != null) {
                 ctx.fillStyle = this.selectionColor;